Карта перемещена на фрагмент, добавлен элемент в список
