+++ app/src/main/java/com/example/infograce/mapStyleModels/TilesetSource.kt
@@ -1,0 +1,6 @@
+package com.example.infograce.mapStyleModels
+
+data class TilesetSource(
+    val id: String,
+    val url: String,
+)
