+++ app/src/main/res/layout/main_activity.xml
@@ -1,34 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/drawer_layout"
     android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
     android:orientation="horizontal"
+    android:layout_height="fill_parent"
     tools:context=".MainActivity">
 
-    <com.mapbox.maps.MapView xmlns:mapbox="http://schemas.android.com/apk/res-auto"
-        android:id="@+id/mapView"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragment1"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        mapbox:mapbox_cameraTargetLat="55.7558"
-        mapbox:mapbox_cameraTargetLng="37.6173"
-        mapbox:mapbox_cameraZoom="14.0" />
+        android:name="com.example.infograce.MainFragment"/>
 
-    <com.google.android.material.navigation.NavigationView
-
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="right"
-        android:layout_marginLeft="-64dp"
-        android:background="@color/dark_blue"
-        android:paddingTop="24dp">
-
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/fragment1"
-            android:name="com.example.infograce.PanelFragment"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-    </com.google.android.material.navigation.NavigationView>
-
-</androidx.drawerlayout.widget.DrawerLayout>
+</LinearLayout>
